requirements.txt
Django==1.4
PIL==1.1.7
boto==2.5.2
dj-database-url==0.2.1
django-discover-runner==0.2
django-storages==1.1.5
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5
wsgiref==0.1.2
